+++ sklearn/input_sample/README.md
@@ -1,9 +1,2 @@
-## CURL examples
-
-### Without transformer
+## CURL example
 > `curl -kv -X POST -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" https://<IP>:32222/dkube/inf/v1/models/d3-inf-3d133119-ec8d-47ce-851f-a7608f4cadd4:predict -d @stock-input.json`
-
-### With transformer
-> `curl -kv -X POST -H "Content-Type: application/json" -H "Authorization: Bearer $TOKEN" https://<IP>:32222/dkube/inf/v1/models/d3-inf-3d133119-ec8d-47ce-851f-a7608f4cadd4:predict -d @stock-input-transformer.json`
-
-Replace the value of *token* with your auth token in `stock-input-transformer.json`
