--- sklearn/input_sample/stock-input-transformer.json
@@ -1,13 +1,0 @@
-{
-    "signatures": {
-        "inputs": [
-            [
-                {
-                    "data": "\ufeffDate,Open,High,Low,Close,Volume\n26-Feb-16,708.58,713.43,700.86,705.07,2239978\n25-Feb-16,700.01,705.98,690.58,705.75,1631855\n24-Feb-16,688.92,700.00,680.78,699.56,1958611\n23-Feb-16,701.45,708.40,693.58,695.85,1999699\n22-Feb-16,707.45,713.24,702.51,706.46,1946067\n19-Feb-16,695.03,703.08,694.05,700.91,1582260\n18-Feb-16,710.00,712.35,696.03,697.35,1859130\n17-Feb-16,699.00,709.75,691.38,708.40,2466808\n16-Feb-16,692.98,698.00,685.05,691.00,2497024\n12-Feb-16,690.26,693.75,678.60,682.40,2129831\n11-Feb-16,675.00,689.35,668.87,683.11,3007223\n10-Feb-16,686.86,701.31,682.13,684.12,2627379\n9-Feb-16,672.32,699.90,668.77,678.11,3604335\n8-Feb-16,667.85,684.03,663.06,682.74,4212541\n5-Feb-16,703.87,703.99,680.15,683.57,5069985\n4-Feb-16,722.81,727.00,701.86,708.01,5145855\n3-Feb-16,770.22,774.50,720.50,726.95,6162333\n2-Feb-16,784.50,789.87,764.65,764.65,6332431\n1-Feb-16,750.46,757.86,743.27,752.00,4801816\n"
-                }
-            ]
-        ]
-    },
-    "instances": [],
-    "token": "dumy token"
-}
