+++ sklearn/input_sample/stock-input.json
@@ -1,19 +1,7 @@
 {
     "instances": [
         [
-            "26"
-        ],
-        [
-            "25"
-        ],
-        [
-            "24"
-        ],
-        [
-            "23"
-        ],
-        [
-            "22"
+            708.58
         ]
     ]
 }
